+++ fin_api_functions/lib/fin_api_functions.dart
@@ -1,3 +1,4 @@
 library fin_api_functions;
 
 export 'src/services/api_functions_service.dart';
+export 'src/models/real_estate_model.dart';